requirements.txt
yfinance
pandas
requests
lxml
black
rich
httpx
git+https://github.com/twopirllc/pandas-ta.git@development
APScheduler
psycopg2-binary
dotenv
textblob
scipy
notebook
matplotlib
seaborn
scikit-learn
tensorflow
xgboost